Ignore button presses after boot, add ATTR_PACKED attribute to HID report structs
