+++ fw/Descriptors.cpp
@@ -7,8 +7,24 @@ const USB_Descriptor_Configuration_t* ConfigurationDescriptor;
 uint8_t LedStringCount;
 const USB_Descriptor_String_t** LedStrings;
 
+// Language descriptor structure. This descriptor, located in FLASH
+// memory, is returned when the host requests the string descriptor
+// with index 0 (the first index). It is actually an array of 16-bit
+// integers, which indicate via the language ID table available at
+// USB.org what languages the device supports for its string
+// descriptors.
 const USB_Descriptor_String_t PROGMEM LanguageString = USB_STRING_DESCRIPTOR_ARRAY(LANGUAGE_ID_ENG);
+
+// Manufacturer descriptor string. This is a Unicode string containing
+// the manufacturer's details in human readable form, and is read out
+// upon request by the host when the appropriate string ID is
+// requested, listed in the Device Descriptor.
 const USB_Descriptor_String_t PROGMEM ManufacturerString = USB_STRING_DESCRIPTOR(L"SEGV");
+
+// Product descriptor string. This is a Unicode string containing the
+// product's details in human readable form, and is read out upon
+// request by the host when the appropriate string ID is requested,
+// listed in the Device Descriptor.
 const USB_Descriptor_String_t PROGMEM ProductString = USB_STRING_DESCRIPTOR(L"BEEF BOARD");
 
 // This function is called by the library when in device mode, and
