+++ fw/Descriptors.h
@@ -72,63 +72,11 @@ HID_RI_COLLECTION(8, 0x02), \
   HID_RI_OUTPUT(8, HID_IOF_DATA | HID_IOF_VARIABLE | HID_IOF_ABSOLUTE | HID_IOF_NON_VOLATILE), \
 HID_RI_END_COLLECTION(0)
 
-#define HID_KEYBOARD_DESCRIPTOR(MaxKeys) \
-HID_RI_USAGE_PAGE(8, 0x01), \
-HID_RI_USAGE(8, 0x06), \
-HID_RI_COLLECTION(8, 0x01), \
-  HID_RI_USAGE_PAGE(8, 0x07), \
-  HID_RI_USAGE_MINIMUM(8, 0x00), \
-  HID_RI_USAGE_MAXIMUM(8, 0xFF), \
-  HID_RI_LOGICAL_MINIMUM(8, 0x00), \
-  HID_RI_LOGICAL_MAXIMUM(16, 0xFF), \
-  HID_RI_REPORT_COUNT(8, MaxKeys), \
-  HID_RI_REPORT_SIZE(8, 0x08), \
-  HID_RI_INPUT(8, HID_IOF_DATA | HID_IOF_ARRAY | HID_IOF_ABSOLUTE), \
-HID_RI_END_COLLECTION(0)
-
-#define HID_MOUSE_DESCRIPTOR \
-HID_RI_USAGE_PAGE(8, 0x01), \
-HID_RI_USAGE(8, 0x02), \
-HID_RI_COLLECTION(8, 0x01), \
-  HID_RI_USAGE(8, 0x01), \
-  HID_RI_COLLECTION(8, 0x02),\
-    HID_RI_USAGE(8, 0x31), \
-    HID_RI_USAGE(8, 0x30), \
-    HID_RI_LOGICAL_MINIMUM(8, -1), \
-    HID_RI_LOGICAL_MAXIMUM(8, 1), \
-    HID_RI_PHYSICAL_MINIMUM(8, -1), \
-    HID_RI_PHYSICAL_MAXIMUM(8, 1), \
-    HID_RI_REPORT_COUNT(8, 0x02), \
-    HID_RI_REPORT_SIZE(8, 0x08), \
-    HID_RI_INPUT(8, HID_IOF_DATA | HID_IOF_VARIABLE | HID_IOF_RELATIVE), \
-  HID_RI_END_COLLECTION(0), \
-HID_RI_END_COLLECTION(0)
-
-// Language descriptor structure. This descriptor, located in FLASH
-// memory, is returned when the host requests the string descriptor
-// with index 0 (the first index). It is actually an array of 16-bit
-// integers, which indicate via the language ID table available at
-// USB.org what languages the device supports for its string
-// descriptors.
-extern const USB_Descriptor_String_t PROGMEM LanguageString;
-
-// Manufacturer descriptor string. This is a Unicode string containing
-// the manufacturer's details in human readable form, and is read out
-// upon request by the host when the appropriate string ID is
-// requested, listed in the Device Descriptor.
-extern const USB_Descriptor_String_t PROGMEM ManufacturerString;
-
-// Product descriptor string. This is a Unicode string containing the
-// product's details in human readable form, and is read out upon
-// request by the host when the appropriate string ID is requested,
-// listed in the Device Descriptor.
-extern const USB_Descriptor_String_t PROGMEM ProductString;
-
 // Type define for the device configuration descriptor structure. This
 // must be defined in the application code, as the configuration
 // descriptor contains several sub-descriptors which vary between
 // devices, and which describe the device's usage to the host.
-
+//
 // Each interface can only support one IN and OUT endpoint,
 // so for KBM support we need to use multiple interfaces as annoying as it is.
 // Technically keyboard data could be sent in the joystick IN endpoint, but might as well do things "cleanly".
@@ -187,11 +135,37 @@ enum {
 };
 
 const USB_Descriptor_HIDReport_Datatype_t PROGMEM KeyboardHIDReport[] = {
-  HID_KEYBOARD_DESCRIPTOR(KEYBOARD_KEYS)
+  HID_RI_USAGE_PAGE(8, 0x01), \
+  HID_RI_USAGE(8, 0x06), \
+  HID_RI_COLLECTION(8, 0x01), \
+    HID_RI_USAGE_PAGE(8, 0x07), \
+    HID_RI_USAGE_MINIMUM(8, 0x00), \
+    HID_RI_USAGE_MAXIMUM(8, 0xFF), \
+    HID_RI_LOGICAL_MINIMUM(8, 0x00), \
+    HID_RI_LOGICAL_MAXIMUM(16, 0xFF), \
+    HID_RI_REPORT_COUNT(8, KEYBOARD_KEYS), \
+    HID_RI_REPORT_SIZE(8, 0x08), \
+    HID_RI_INPUT(8, HID_IOF_DATA | HID_IOF_ARRAY | HID_IOF_ABSOLUTE), \
+  HID_RI_END_COLLECTION(0)
 };
 
 const USB_Descriptor_HIDReport_Datatype_t PROGMEM MouseHIDReport[] {
-  HID_MOUSE_DESCRIPTOR
+  HID_RI_USAGE_PAGE(8, 0x01), \
+  HID_RI_USAGE(8, 0x02), \
+  HID_RI_COLLECTION(8, 0x01), \
+    HID_RI_USAGE(8, 0x01), \
+    HID_RI_COLLECTION(8, 0x02),\
+      HID_RI_USAGE(8, 0x31), \
+      HID_RI_USAGE(8, 0x30), \
+      HID_RI_LOGICAL_MINIMUM(8, -1), \
+      HID_RI_LOGICAL_MAXIMUM(8, 1), \
+      HID_RI_PHYSICAL_MINIMUM(8, -1), \
+      HID_RI_PHYSICAL_MAXIMUM(8, 1), \
+      HID_RI_REPORT_COUNT(8, 0x02), \
+      HID_RI_REPORT_SIZE(8, 0x08), \
+      HID_RI_INPUT(8, HID_IOF_DATA | HID_IOF_VARIABLE | HID_IOF_RELATIVE), \
+    HID_RI_END_COLLECTION(0), \
+  HID_RI_END_COLLECTION(0)
 };
 
 // Device descriptor structure. This descriptor, located in FLASH
