+++ fw/combo.cpp
@@ -1,8 +1,7 @@
-#include "devices/iidx/iidx_rgb_manager.h"
-
 #include "beef.h"
 #include "combo.h"
 #include "hid.h"
+#include "rgb_helper.h"
 
 combo (*get_button_combo_callback) (uint16_t);
 
@@ -54,7 +53,7 @@ void process_combos(config* current_config,
 
     timer_init(&combo_timer);
     timer_init(combo_lights_timer);
-    timer_init(&IIDX::RgbManager::Turntable::combo_timer);
+    timer_init(&RgbHelper::combo_timer);
 
     if (config_update_callback.addr != nullptr)
       config_update(config_update_callback.addr, config_update_callback.val);
