--- fw/devices/iidx/iidx_rgb.h
@@ -1,13 +1,0 @@
-#pragma once
-
-#include <stdint.h>
-
-#include "../rgb.h"
-
-namespace IIDX {
-  struct hid_lights {
-    uint16_t buttons;
-    rgb_light tt_lights;
-    rgb_light bar_lights;
-  };
-}
