--- fw/devices/sdvx/sdvx_rgb.h
@@ -1,9 +1,0 @@
-#pragma once
-
-#include <stdint.h>
-
-namespace SDVX {
-  struct hid_lights {
-    uint16_t buttons;
-  };
-}
