--- fw/devices/sdvx/sdvx_rgb_manager.cpp
@@ -1,6 +1,0 @@
-#include "sdvx_rgb_manager.h"
-
-namespace SDVX {
-  namespace RgbManager {
-  }
-}
