+++ fw/rgb_helper.cpp
@@ -11,7 +11,11 @@ namespace RgbHelper {
     TT_DATA_PIN  = 15  // C5
   };
 
+  timer combo_timer{};
+
   void init() {
+    timer_init(&combo_timer);
+
     FastLED.addLeds<NEOPIXEL, TT_DATA_PIN>(tt_leds, RING_LIGHT_LEDS)
       .setDither(DISABLE_DITHER);
     FastLED.addLeds<NEOPIXEL, BAR_DATA_PIN>(bar_leds, LIGHT_BAR_LEDS)
